fix linter

+++ core/txpool/txpool.go
@@ -1898,7 +1898,7 @@ func (pool *TxPool) updateBaseFeeAt(head *types.Header) error {
 	if err != nil {
 		return err
 	}
-	_, baseFeeEstimate, err := dummy.EstimateNextBaseFee(pool.chainconfig, feeConfig, head, uint64(time.Now().Unix()))
+	_, baseFeeEstimate, err := dummy.EstimateNextBaseFee(pool.chainconfig, feeConfig, head, time.Now().Unix())
 	if err != nil {
 		return err
 	}
